Added Mensia database to Teacher Component.
This commit allow the lookup of a Teacher from the Mensiaia Database.
Currently only using the integer ID, but will be moved to use a UUID in
the near future.

Seed data will need to be added to the database before the find works.
This will be added as a mix task.

To create the database a new mix task has been made:

mix teacher_component.setup